backend/portfolio-upload/requirements.txt
boto3==1.35.82
psycopg2-binary==2.9.10
replicate==0.34.1
requests==2.32.3
